+++ src/app/common/validation/forms/currency-validators.ts
@@ -1,4 +1,3 @@
-import BigNumber from 'bignumber.js';
 import * as yup from 'yup';
 
 import { BTC_DECIMALS, STX_DECIMALS } from '@shared/constants';
@@ -36,29 +35,6 @@ function currencyPrecisionValidatorFactory(
 
 export function btcAmountPrecisionValidator(errorMsg: string) {
   return currencyPrecisionValidatorFactory('BTC', BTC_DECIMALS, errorMsg);
-}
-
-interface BtcInsufficientBalanceValidatorArgs {
-  recipient: string;
-  calcMaxSpend(recipient: string): {
-    spendableBitcoin: BigNumber;
-  };
-}
-export function btcInsufficientBalanceValidator({
-  recipient,
-  calcMaxSpend,
-}: BtcInsufficientBalanceValidatorArgs) {
-  return yup.number().test({
-    message: 'Insufficient funds',
-    test(value) {
-      if (!value) return false;
-      const maxSpend = calcMaxSpend(recipient);
-      if (!maxSpend) return false;
-      const desiredSpend = new BigNumber(value);
-      if (desiredSpend.isGreaterThan(maxSpend.spendableBitcoin)) return false;
-      return true;
-    },
-  });
 }
 
 export function stxAmountPrecisionValidator(errorMsg: string) {
