+++ src/app/pages/send/send-crypto-asset-form/form/btc/use-btc-send-form.tsx
@@ -15,8 +15,11 @@ import {
   btcAddressNetworkValidator,
   btcAddressValidator,
 } from '@app/common/validation/forms/address-validators';
+import {
+  btcInsufficientBalanceValidator,
+  btcMinimumSpendValidator,
+} from '@app/common/validation/forms/amount-validators';
 import { btcAmountPrecisionValidator } from '@app/common/validation/forms/currency-validators';
-import { btcInsufficientBalanceValidator } from '@app/common/validation/forms/currency-validators';
 import { useBitcoinAssetBalance } from '@app/query/bitcoin/address/address.hooks';
 import { useCurrentBtcNativeSegwitAccountAddressIndexZero } from '@app/store/accounts/blockchain/bitcoin/native-segwit-account.hooks';
 import { useCurrentNetwork } from '@app/store/networks/networks.selectors';
@@ -53,7 +56,8 @@ export function useBtcSendForm() {
             recipient: formRef.current?.values.recipient ?? '',
             calcMaxSpend,
           })
-        ),
+        )
+        .concat(btcMinimumSpendValidator()),
       recipient: yup
         .string()
         .concat(btcAddressValidator())
